restapi/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-bcrypt
flask-restful
flask-jwt-extended
flask-cors
requests-oauthlib
marshmallow
psycopg2-binary
python-dotenv
python-slugify
Pillow
pytest
flake8
redis
